Including links back in event struct to allow extracting of ack id
Hopefully fixes #1

Otherwise we probably have to ack and nack by doing 1 req per event,
or change the API of read_from_subscription to include the actual
Response struct (where the ackAll relation link is present) which
ofc would be more "correct" but I don't like that we have to expose
the Response struct as subscriptions are higher level imo.
Anywho, that is the alternative if this fix doesn't cut it.